refactor: delta long으로 통일

+++ src/main/java/com/algangi/mongle/comment/domain/model/Comment.java
@@ -104,12 +104,12 @@ public class Comment extends TimeBaseEntity implements CursorConvertible {
         this.deletedAt = LocalDateTime.now();
     }
 
-    public void increaseLikeCount(int delta) {
+    public void increaseLikeCount(long delta) {
         this.likeCount += delta;
         if (this.likeCount < 0) this.likeCount = 0;
     }
 
-    public void increaseDislikeCount(int delta) {
+    public void increaseDislikeCount(long delta) {
         this.dislikeCount += delta;
         if (this.dislikeCount < 0) this.dislikeCount = 0;
     }
